third_party/blink/web_tests/http/tests/security/upgrade-insecure-requests/https-header-subresource.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    if (window.testRunner)
      testRunner.setPopupBlockingEnabled(false);

    var t = async_test("Verify that subresources do not receive an 'HTTPS' header.");
    fetch("resources/echo-https-header.pl").then(t.step_func(function(res) {
      return res.json();
    })).then(t.step_func(function(json) {
      assert_equals(json.header, "");
      t.done();
    }));
  </script>
</head>
<body>
</body>
</html>
